new: usr: New data splitter sink
A new sink has been included as part of the default sinks of flowbber.
The DataSplitterSink allows user to extract pieces of data and create
files with that data. Check the docs to see how to use them. The valgrind
example contains a basic usage of how this sink works.